Fixed packages: classes separations on the packages.
