--- src/main/java/BurgersResponse.java
@@ -1,10 +1,0 @@
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class BurgersResponse {
-    public int statusCode;
-    public String message;
-    public String accessToken;
-}
