+++ src/main/java/burgers/api/BurgersAPISpecifications.java
@@ -1,3 +1,5 @@
+package burgers.api;
+
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.http.ContentType;
 import io.restassured.specification.RequestSpecification;
