+++ src/main/java/burgers/models/BurgersOrder.java
@@ -1,3 +1,5 @@
+package burgers.models;
+
 import lombok.Builder;
 import lombok.Data;
 
@@ -5,12 +7,14 @@ import java.util.List;
 
 @Data
 @Builder
-public class BurgersGetOrder {
+public class BurgersOrder {
+    private List<BurgersIngredient> ingredients;
     private String _id;
-    private List<String> ingredients;
+    private BurgersOrderOwner owner;
     private String status;
     private String name;
     private String createdAt;
     private String updatedAt;
     private int number;
+    private int price;
 }
